+++ test/chunk/soc.spec.ts
@@ -1,9 +1,8 @@
 import { Bytes, verifyBytes } from '../../src/utils/bytes'
-import { makeSingleOwnerChunk, verifySingleOwnerChunk } from '../../src/chunk/soc'
+import { makeSingleOwnerChunk, verifySingleOwnerChunk, uploadSingleOwnerChunk } from '../../src/chunk/soc'
 import { makeContentAddressedChunk, verifyChunk } from '../../src/chunk/cac'
 import { beeUrl, testIdentity } from '../utils'
 import { makeDefaultSigner } from '../../src/chunk/signer'
-import { uploadSingleOwnerChunk } from '../../src/chunk/upload'
 import { serializeBytes } from '../../src/chunk/serialize'
 import { makeSpan } from '../../src/chunk/span'
 import * as chunkAPI from '../../src/modules/chunk'
